+++ Cpp/Basic Problems/Chaining
