+++ Cpp/Basic Problems/Knuth-Morris-Pratt Algorithm
