+++ Cpp/Basic Problems/Kruskals_minspanning_tree_function
